Separate context and provider
