+++ client/src/index.js
@@ -4,7 +4,7 @@ import ReactDOM from 'react-dom'
 import 'tachyons/css/tachyons.min.css'
 import { Provider } from 'unstated'
 import App from './App.js'
-import { AppContextProvider } from './containers/AppContext'
+import AppContextProvider from './containers/AppContextProvider'
 import './css/index.css'
 import './css/react-split-pane.css'
 import './css/vendorOverrides.css'
